playwright_ui_testing/pyproject.toml
[project]
name = "playwright-ui-testing"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-playwright (>=0.7.0,<0.8.0)",
    "pytest-bdd (>=8.1.0,<9.0.0)"
]

[tool.poetry]
packages = [{include = "playwright_ui_testing", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = 6.0
addopts = "-ra -q --cucumber-json=test.json --tracing=retain-on-failure"
testpaths = "tests/"
python_files = "step_*.py"